requirements.txt
fastapi>=0.109.0
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic
openai>=1.72.0
google-generativeai

# Agent dependencies
mcp[cli]>=1.6.0

# Testing dependencies
pytest==7.4.0
pytest-asyncio==0.21.1
httpx
requests>=2.32.3  # Updated to meet smolagents requirement
